Promote core and fortigate-autoscale to packages, use npm to link them together.
