+++ core/autoscale-environment.ts
