+++ core/blob.ts
