+++ core/cloud-function-proxy.ts
