+++ core/context-strategy/licensing-context.ts
