+++ core/db-definitions.ts
