+++ core/helper-function.ts
