+++ core/index.ts
@@ -15,5 +15,6 @@ export * from './platform-adaptee';
 export * from './platform-adapter';
 export * from './primary-election';
 export * from './virtual-machine';
+export * from './faz-integration-strategy';
 // NOTE: the index will not export the db-definitions due to name conflicts.
 // Import them directly from the ./db-definitions.ts when needed.
