+++ core/platform-adaptee.ts
