+++ core/platform-adapter.ts
