+++ core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../tsconfig-base.json",
+    "compilerOptions": {
+        "outDir": "dist",
+        "rootDir": "."
+    } //,
+    //"files": ["index.ts"]
+}
