+++ core/virtual-machine.ts
