+++ package.json
@@ -8,9 +8,9 @@
         "pre-push": "npm-run-all lint-check audit-production test build",
         "audit-production": "npm audit --production",
         "bundle": "webpack",
-        "compile": "tsc",
-        "clear-dist": "rimraf dist",
-        "build": "npm run clear-dist && npm run compile",
+        "compile": "tsc --build",
+        "clear-dist": "rimraf */dist",
+        "build": "cd fortigate-autoscale; npm install; cd ..; npm run clear-dist && npm run compile",
         "make-dist": "node dist/scripts/make-dist.js",
         "build-artifacts": "npm-run-all build make-dist",
         "test": "mocha --timeout 3600000 -r ts-node/register -r tsconfig-paths/register test/{*,**/**/*}.spec.ts",
